refactor: adapt to patcher changes
