+++ src/main/kotlin/app/revanced/cli/command/MainCommand.kt
@@ -10,7 +10,7 @@ import app.revanced.patcher.PatcherOptions
 import app.revanced.patcher.extensions.PatchExtensions.compatiblePackages
 import app.revanced.patcher.extensions.PatchExtensions.description
 import app.revanced.patcher.extensions.PatchExtensions.patchName
-import app.revanced.patcher.util.patch.implementation.JarPatchBundle
+import app.revanced.patcher.util.patch.impl.JarPatchBundle
 import app.revanced.utils.adb.Adb
 import picocli.CommandLine.*
 import java.io.File
